// web/src/typings/slot.ts
export type Slot = {
  slot: number;
  name?: string;
  count?: number;
  weight?: number;
  metadata?: {
    [key: string]: any;
  };
  durability?: number;
};

export type SlotWithItem = Slot & {
  name: string;
  count: number;
  weight: number;
  durability?: number;
  price?: number;
  currency?: string;
  ingredients?: { [key: string]: number };
  duration?: number;
  image?: string;
  grade?: number | number[];
  // ADD THESE TWO LINES:
  inventoryKey?: 'leftInventory' | 'backpackInventory' | 'rightInventory';
  inventoryType?: string;
};